requirements.txt
python-dotenv
pandas
tiktoken
openai>=1.0.0
textstat
regex
fastapi
uvicorn
streamlit
pydantic
google-generativeai
pyyaml
pymupdf
